+++ test/cookbooks/test/metadata.rb
@@ -8,3 +8,4 @@ version           '0.0.1'
 
 depends           'apache2'
 depends           'apt'
+depends           'yum-ius'