+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of the apache2 cookbook.
 ## Unreleased
 
 - resolved cookstyle error: test/cookbooks/test/recipes/php.rb:1:1 refactor: `ChefCorrectness/IncorrectLibraryInjection`
+- Allow override of package name and version in `install` resource
 
 ## 8.3.0 (2020-07-13)
 
@@ -78,7 +79,6 @@ This file is used to list changes made in each version of the apache2 cookbook.
 - Allow users to alter the default modules configuration without re### Added
 
 - Uniform way to pass IP's in mod_info and mod_status
-- Allow override of package name and version in `install` resource
 
 ## [7.0.0] - 05-03-2019